refactor: split locale provider and stub payments
