+++ components/Footer.tsx
@@ -6,7 +6,7 @@ import { SocialIconLink } from '@/components/design-system/SocialIconLink';
 import { MailIcon, PhoneIcon, MapPinIcon, ClockIcon } from '@/components/design-system/icons';
 import { FooterGrid } from '@/components/design-system/FooterGrid';
 import { FooterSection } from './FooterSection';
-import { useLocale } from '@/lib/locale';
+import { useLocale } from '@/lib/localeProvider';
 
 import {
   resources,
